Move email address related methods to GitHub
Add integration test for adding email addresses

Remove integration tests for User#email_addresses
