+++ tests/cassettes/GitHub_add_email_addresses.json
@@ -1,0 +1,1 @@
+{"recorded_with": "betamax/0.5.1", "http_interactions": [{"recorded_at": "2015-12-20T18:56:05", "response": {"url": "https://api.github.com/user/emails", "status": {"code": 201, "message": "Created"}, "headers": {"Server": "GitHub.com", "Date": "Sun, 20 Dec 2015 18:56:05 GMT", "X-GitHub-Media-Type": "github.v3; param=full; format=json", "ETag": "\"b4fdbf10626c0cffe2eb8fe3ad230ec1\"", "X-Content-Type-Options": "nosniff", "X-RateLimit-Limit": "5000", "Content-Security-Policy": "default-src 'none'", "Content-Type": "application/json; charset=utf-8", "X-GitHub-Request-Id": "451DE1EB:14B57:380B98F:5676F9C5", "Content-Length": "206", "Access-Control-Allow-Credentials": "true", "X-RateLimit-Remaining": "4997", "Access-Control-Expose-Headers": "ETag, Link, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval", "Cache-Control": "private, max-age=60, s-maxage=60", "Access-Control-Allow-Origin": "*", "Vary": "Accept, Authorization, Cookie, X-GitHub-OTP", "X-RateLimit-Reset": "1450639603", "Strict-Transport-Security": "max-age=31536000; includeSubdomains; preload", "X-Served-By": "2c18a09f3ac5e4dd1e004af7c5a94769", "X-Frame-Options": "deny", "X-XSS-Protection": "1; mode=block", "Status": "201 Created"}, "body": {"encoding": "utf-8", "string": "[{\"email\":\"gr.aff.at.colmingov@gmail.com\",\"primary\":true,\"verified\":true},{\"email\":\"example1@example.com\",\"primary\":false,\"verified\":false},{\"email\":\"example2@example.com\",\"primary\":false,\"verified\":false}]"}}, "request": {"method": "POST", "uri": "https://api.github.com/user/emails", "headers": {"Authorization": "Basic <BASIC_AUTH>", "Content-Type": "application/json", "Accept": "application/vnd.github.v3.full+json", "Content-Length": "48", "Accept-Charset": "utf-8", "Accept-Encoding": "gzip, deflate", "User-Agent": "github3.py/1.0.0a2", "Connection": "keep-alive"}, "body": {"encoding": "utf-8", "string": "[\"example1@example.com\", \"example2@example.com\"]"}}}]}
