--- tests/cassettes/User_email_addresses.json
@@ -1,1 +1,0 @@
-{"http_interactions": [{"response": {"headers": {"Server": "GitHub.com", "X-Served-By": "b0ef53392caa42315c6206737946d931", "X-Content-Type-Options": "nosniff", "X-GitHub-Request-Id": "4B80D025:1D25C:4C6CA87:5674605C", "Transfer-Encoding": "chunked", "X-Accepted-OAuth-Scopes": "", "Access-Control-Expose-Headers": "ETag, Link, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval", "Content-Security-Policy": "default-src 'none'", "X-RateLimit-Limit": "5000", "Strict-Transport-Security": "max-age=31536000; includeSubdomains; preload", "Access-Control-Allow-Origin": "*", "Content-Encoding": "gzip", "Status": "200 OK", "X-XSS-Protection": "1; mode=block", "X-RateLimit-Remaining": "4985", "X-Frame-Options": "deny", "X-OAuth-Scopes": "gist, repo, user", "Date": "Fri, 18 Dec 2015 19:37:00 GMT", "ETag": "W/\"331e4edb3c6861ab1d3aeb7f0bfa0a7e\"", "X-GitHub-Media-Type": "github.v3; param=full; format=json", "Vary": "Accept, Authorization, Cookie, X-GitHub-OTP", "Access-Control-Allow-Credentials": "true", "X-RateLimit-Reset": "1450468914", "Last-Modified": "Thu, 17 Dec 2015 00:57:45 GMT", "Content-Type": "application/json; charset=utf-8", "Cache-Control": "private, max-age=60, s-maxage=60"}, "url": "https://api.github.com/user", "status": {"message": "OK", "code": 200}, "body": {"encoding": "utf-8", "base64_string": "H4sIAAAAAAAAA51UTY+bMBT8KyufyfIRkmyQqlbqaQ+9VFtV6iV6GIdYa2xkG6IUbX97x0CqLj1UySnk8WbeeJ6HgSlTS80K5mTdUC9t57KcRUxWrMjy5GmdRIx68mQPnVXoO3nfuiKOp6J7rKU/dWXnhOVGe6H9IzdN3MUT+GP/YQ222s4cgZahsOBq5cwzgUHm4oWgk2/UQsI0eoQsmo9GKXMGy1L0/wbFf5AQOT1LXd/JAuQQG38S8A5HegtGSOdvFzWihjj8HGQVeBwWYkV1s7AZB1lnDUVDbEVrRsKudNzK1kujbxf4Dg02Y2vS8ifdxwa0A0mQdruUEQW06HEXb4dPsCFureyJX4I1VnAhe5h9J+UCD0Z/aQVy8A2XIlgvvThQ1YQYHkk5ETFNTWh4Jv3w2diKHDdoxEVvSV/w4ivxV9cSF6iWSPCcS8TyfD4jELWS5ZiLX5JP+HhsixB3Pm4FiC9USWd09PD9OfjVkAz5RlSPR/LcKAiqTf8JnwWpAhmaTtIKKhW06U4pzJbm+th2pZL8MNlfpNk6YnNpvL2seMquiUIu8XHZbv9KWPiPA4Ldw2jyEJIlabJK8lWavCR5keyKdfYDErq2etezWaXZKt29JEmx2RX5JvSMy4Op8+Q0xWhvPKnD9c0sE/UQhGpZTyMGb14RW6px2PV2F76D8ERRaSx5M54AgxTByeG6r0ZyGxY1rabY77abbZbv9/9AQbZQsnl7+w2ejadljAUAAA==", "string": ""}}, "recorded_at": "2015-12-18T19:37:00", "request": {"headers": {"Connection": "keep-alive", "User-Agent": "github3.py/1.0.0a2", "Accept": "application/vnd.github.v3.full+json", "Accept-Charset": "utf-8", "Content-Type": "application/json", "Accept-Encoding": "gzip, deflate", "Authorization": "token <AUTH_TOKEN>"}, "uri": "https://api.github.com/user", "method": "GET", "body": {"encoding": "utf-8", "string": ""}}}, {"response": {"headers": {"Server": "GitHub.com", "X-Served-By": "3e3b9690823fb031da84658eb58aa83b", "X-Content-Type-Options": "nosniff", "X-GitHub-Request-Id": "4B80D025:1D25C:4C6CA99:5674605C", "Transfer-Encoding": "chunked", "X-Accepted-OAuth-Scopes": "user, user:email", "Access-Control-Expose-Headers": "ETag, Link, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval", "Content-Security-Policy": "default-src 'none'", "X-RateLimit-Limit": "5000", "Strict-Transport-Security": "max-age=31536000; includeSubdomains; preload", "Access-Control-Allow-Origin": "*", "Content-Encoding": "gzip", "Status": "200 OK", "X-GitHub-Media-Type": "github.v3; param=full; format=json", "X-RateLimit-Remaining": "4984", "X-Frame-Options": "deny", "X-OAuth-Scopes": "gist, repo, user", "Date": "Fri, 18 Dec 2015 19:37:00 GMT", "ETag": "W/\"f2a8db3c8bcf67d6a066b9b77316d886\"", "X-XSS-Protection": "1; mode=block", "Vary": "Accept, Authorization, Cookie, X-GitHub-OTP", "Access-Control-Allow-Credentials": "true", "X-RateLimit-Reset": "1450468914", "Content-Type": "application/json; charset=utf-8", "Cache-Control": "private, max-age=60, s-maxage=60"}, "url": "https://api.github.com/user/emails?per_page=100", "status": {"message": "OK", "code": 200}, "body": {"encoding": "utf-8", "base64_string": "H4sIAAAAAAAAA53QMQ7CMAyF4btkpcqAmJhyD8Tgpk6wmsSVkxRViLsT6IDEAmX08D5Z/+mmMAIFdVRkWQbIduepXGpvLPtAvbYcVacmoQiyqGORip2aUcgRDut9796KF3AOiuUQKXmejX/yn4qDkL8zGkobBt0k/T9FkEyPaViuLGPe/Embt82rDBsBO+YJLG5mMrUQM0nN+4OpGSXrxIJTWPTa+zfx/AAJ1R1ksQEAAA==", "string": ""}}, "recorded_at": "2015-12-18T19:37:00", "request": {"headers": {"Connection": "keep-alive", "User-Agent": "github3.py/1.0.0a2", "Accept": "application/vnd.github.v3.full+json", "Accept-Charset": "utf-8", "Content-Type": "application/json", "Accept-Encoding": "gzip, deflate", "Authorization": "token <AUTH_TOKEN>"}, "uri": "https://api.github.com/user/emails?per_page=100", "method": "GET", "body": {"encoding": "utf-8", "string": ""}}}], "recorded_with": "betamax/0.5.1"}
